Renamed to token classification
