+++ exasol_transformers_extension/udfs/callers/token_classification_udf_call.py
@@ -1,0 +1,8 @@
+from exasol_transformers_extension.udfs.models.token_classification_udf import \
+    TokenClassificationUDF
+
+udf = TokenClassificationUDF(exa)
+
+
+def run(ctx):
+    return udf.run(ctx)
