+++ tests/unit_tests/udfs/test_token_classification.py
@@ -4,31 +4,31 @@ from exasol_udf_mock_python.group import Group
 from exasol_udf_mock_python.mock_exa_environment import MockExaEnvironment
 from exasol_udf_mock_python.mock_meta_data import MockMetaData
 from exasol_udf_mock_python.udf_mock_executor import UDFMockExecutor
-from tests.unit_tests.udf_wrapper_params.named_entity_recognition.multiple_bfsconn_single_subdir_single_model_multiple_batch import \
+from tests.unit_tests.udf_wrapper_params.token_classification.multiple_bfsconn_single_subdir_single_model_multiple_batch import \
     MultipleBucketFSConnSingleSubdirSingleModelNameMultipleBatch
-from tests.unit_tests.udf_wrapper_params.named_entity_recognition.multiple_bfsconn_single_subdir_single_model_single_batch import \
+from tests.unit_tests.udf_wrapper_params.token_classification.multiple_bfsconn_single_subdir_single_model_single_batch import \
     MultipleBucketFSConnSingleSubdirSingleModelNameSingleBatch
-from tests.unit_tests.udf_wrapper_params.named_entity_recognition.multiple_model_multiple_batch_complete import \
+from tests.unit_tests.udf_wrapper_params.token_classification.multiple_model_multiple_batch_complete import \
     MultipleModelMultipleBatchComplete
-from tests.unit_tests.udf_wrapper_params.named_entity_recognition.multiple_model_multiple_batch_incomplete import \
+from tests.unit_tests.udf_wrapper_params.token_classification.multiple_model_multiple_batch_incomplete import \
     MultipleModelMultipleBatchIncomplete
-from tests.unit_tests.udf_wrapper_params.named_entity_recognition.multiple_model_multiple_batch_multiple_models_per_batch import \
+from tests.unit_tests.udf_wrapper_params.token_classification.multiple_model_multiple_batch_multiple_models_per_batch import \
     MultipleModelMultipleBatchMultipleModelsPerBatch
-from tests.unit_tests.udf_wrapper_params.named_entity_recognition.multiple_model_single_batch_complete import \
+from tests.unit_tests.udf_wrapper_params.token_classification.multiple_model_single_batch_complete import \
     MultipleModelSingleBatchComplete
-from tests.unit_tests.udf_wrapper_params.named_entity_recognition.multiple_model_single_batch_incomplete import \
+from tests.unit_tests.udf_wrapper_params.token_classification.multiple_model_single_batch_incomplete import \
     MultipleModelSingleBatchIncomplete
-from tests.unit_tests.udf_wrapper_params.named_entity_recognition.single_bfsconn_multiple_subdir_single_model_multiple_batch import \
+from tests.unit_tests.udf_wrapper_params.token_classification.single_bfsconn_multiple_subdir_single_model_multiple_batch import \
     SingleBucketFSConnMultipleSubdirSingleModelNameMultipleBatch
-from tests.unit_tests.udf_wrapper_params.named_entity_recognition.single_bfsconn_multiple_subdir_single_model_single_batch import \
+from tests.unit_tests.udf_wrapper_params.token_classification.single_bfsconn_multiple_subdir_single_model_single_batch import \
     SingleBucketFSConnMultipleSubdirSingleModelNameSingleBatch
-from tests.unit_tests.udf_wrapper_params.named_entity_recognition.single_model_multiple_batch_complete import \
+from tests.unit_tests.udf_wrapper_params.token_classification.single_model_multiple_batch_complete import \
     SingleModelMultipleBatchComplete
-from tests.unit_tests.udf_wrapper_params.named_entity_recognition.single_model_multiple_batch_incomplete import \
+from tests.unit_tests.udf_wrapper_params.token_classification.single_model_multiple_batch_incomplete import \
     SingleModelMultipleBatchIncomplete
-from tests.unit_tests.udf_wrapper_params.named_entity_recognition.single_model_single_batch_complete import \
+from tests.unit_tests.udf_wrapper_params.token_classification.single_model_single_batch_complete import \
     SingleModelSingleBatchComplete
-from tests.unit_tests.udf_wrapper_params.named_entity_recognition.single_model_single_batch_incomplete import \
+from tests.unit_tests.udf_wrapper_params.token_classification.single_model_single_batch_incomplete import \
     SingleModelSingleBatchIncomplete
