+++ tests/utils/db_queries.py
@@ -6,7 +6,7 @@ deployed_script_list = [
     "TE_QUESTION_ANSWERING_UDF",
     "TE_FILLING_MASK_UDF",
     "TE_TEXT_GENERATION_UDF",
-    "TE_NAMED_ENTITY_RECOGNITION_UDF"
+    "TE_TOKEN_CLASSIFICATION_UDF"
 ]
